requirements.txt
asgiref==3.8.1
Django==5.0.3
django-filter==25.1
djangorestframework==3.15.0
drf-yasg==1.21.9
Faker==36.1.1
inflection==0.5.1
Markdown==3.7
packaging==24.2
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
validate-docbr==1.10.0
